feat: shorten mod link names as much as possible
Resolves #9
Resolves #8

See ValveSoftware/steam-for-linux#5753

Steam seems to have issues with long launch arguments when calling it
via `steam -applaunch GAMEID ARGUMENTS`. This results in the client not
responding to any `-applaunch` attempts anymore, and it needs to be
restarted.

When trying to join servers with lots of mods, the `-mod=...` argument
can become too long for Steam to handle, so mod directories should not
be named by their mod ID and mod name and they should instead be kept as
short as possible. This comes at the cost of not being able to identify
mod directories easily anymore, but it's better than not being able to
join certain servers at all and having to restart Steam.

- Remove mod names from mod directory links
- Turn decimal mod IDs into ASCII char strings and convert to base64